Fixed compilation under MacOS X
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@2030 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ applications/osmo4_wx/wxOsmo4.cpp
@@ -158,7 +158,7 @@ u32 get_sys_col(int idx)
 }
 #endif
 
-static void wxOsmo4_progress_cbk(void *usr, char *title, u32 done, u32 total)
+static void wxOsmo4_progress_cbk(void *usr, char *title, u64 done, u64 total)
 {
 	if (!total) return;
 	wxOsmo4Frame *app = (wxOsmo4Frame *)usr;
